requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
groq>=0.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
google-play-scraper==1.2.6
emoji==2.8.0
langdetect==1.0.9
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
